sign up page: styles.css, imports, & primaryLayout
