+++ client/src/app/layouts/PrimaryLayout.js
@@ -9,7 +9,7 @@ import BookGrill from "../pages/BookGrill/BookGrill";
 import SingleGrill from "../pages/SingleGrill/SingleGrill";
 import HowToGrill from "../pages/HowToGrill/HowToGrill";
 import Pricing from "../pages/Pricing/Pricing";
-import SignUp from "../pages/SignUp";
+import SignUp from "../pages/SignUp/SignUp";
 import ThankYouUser from "../pages/ThankYouUser";
 import NoMatch from "../pages/NoMatch";
