+++ client/src/app/pages/SignUp/styles.css
@@ -1,0 +1,18 @@
+.su-main-container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  margin-bottom: 50px;
+}
+
+.su-logo {
+  margin: 40px 0 20px 0;
+  max-width: 100%;
+}
+
+.su-terms {
+  font-size: 13px;
+  color: black;
+  text-align: center;
+}
